requirements.txt
matplotlib==3.4.3
Pillow==8.4.0
torch==1.9.1
torchvision==0.2.2
tqdm==4.62.3
visdom==0.1.8.9
